refactor(guard): extract protect approval helpers
